+++ src/main/resources/allure.properties
@@ -1,0 +1,1 @@
+allure.results.directory=target/allure-results
